+++ fundamentals 1/main.js
@@ -1,1 +1,18 @@
 alert("Hello world! from external script");
+
+// if you want to reassign the values you can use let keyword.
+let age = 20;
+age = 25;
+
+// const birthDate;  //'const' declarations must be initialized.
+const birthYear = 1945;
+
+var job = "programmer"; 
+job = "developer"; 
+
+console.log(a); // undefined (hoisted)
+var a = 5;
+
+//console.log(b); // Error: Cannot access 'b' before initialization
+let b = 10;
+