+++ changelog.md
@@ -1,1 +1,2 @@
 1.  linking a js file.
+2.  Added let, const and var keywords.